Check if packagePath exists before running installer
Right now it's possible for the package path to not exist, but be passed in anyway. One way to set up application as "assistedInstaller: true, perMachine: true". Then upon the elevate.exe being called on the first attempt at updating, to press "No". Then restart the application and press "Yes" on the next elevate when updating.

If you check the path of the package, there isn't a package, so the installer tries to run on an empty file, but fails causing an uninstall for the user.